requirements.txt
Flask==3.0.0
openai==0.27.9
python-dotenv==1.0.0
tiktoken